VelikaAukcija/requirements.txt
Django>=5.1.2,<6.0
djangorestframework>=3.15.2,<4.0
psycopg>=3.1.9
gunicorn>=23.0.0
django-cors-headers>=4.5.0
pytz
asgiref
python-dotenv
sqlparse
requests
google-auth
Pillow